Votes now only need to start with yea, nay, or abs.
This will allow people, for the sake of the logs, to write qualifiers.
(e.g. "yea, but as 30 minutes"). No functional meeting, but some people
(me!) can be passionate about the record.

But the shortcuts "y", "n", and "a" still must be the whole message (in
order to avoid accidental votes).